save chain api url and jwt to db
- chain api: create group result include chain api url and jwt
- nodesdk: save chain api url and jwt to local db when joining group
- nodesdk: GET /nodesdk_api/v1/group/:group_id/apihosts return chain api
  url and jwt
- remove hardcode jwt
- fix GitCommit and ReleaseVersion
