+++ cmd/relaynode.go
@@ -10,6 +10,7 @@ import (
 	"github.com/rumsystem/quorum/internal/pkg/cli"
 	"github.com/rumsystem/quorum/internal/pkg/conn/p2p"
 	"github.com/rumsystem/quorum/internal/pkg/options"
+	"github.com/rumsystem/quorum/internal/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -51,7 +52,7 @@ func runRelaynode(config cli.RelayNodeFlag) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	logger.Infof("Version: %s", GitCommit)
+	logger.Infof("Version: %s", utils.GitCommit)
 	peername := config.PeerName
 
 	relayNodeOpt, err := options.InitRelayNodeOptions(config.ConfigDir, peername)
