+++ cmd/version.go
@@ -3,6 +3,7 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/rumsystem/quorum/internal/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -11,7 +12,7 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Show version",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Printf("%s - %s", ReleaseVersion, GitCommit)
+		fmt.Printf("%s - %s", utils.ReleaseVersion, utils.GitCommit)
 	},
 }
