+++ pkg/chainapi/api/seed.go
@@ -1,26 +1,30 @@
 package api
 
 import (
-	"github.com/labstack/echo/v4"
-	"github.com/rumsystem/quorum/pkg/chainapi/handlers"
 	"net/http"
+
+	"github.com/labstack/echo/v4"
+	rumerrors "github.com/rumsystem/quorum/internal/pkg/errors"
+	"github.com/rumsystem/quorum/internal/pkg/utils"
+	"github.com/rumsystem/quorum/pkg/chainapi/handlers"
 )
+
+type SeedUrlextendParam struct {
+	SeedURL string `json:"seed" validate:"required"`
+}
 
 func (h *Handler) SeedUrlextend() echo.HandlerFunc {
 	return func(c echo.Context) error {
-		var err error
-		output := make(map[string]string)
-		input := make(map[string]string)
+		cc := c.(*utils.CustomContext)
 
-		if err = c.Bind(&input); err != nil {
-			output[ERROR_INFO] = err.Error()
-			return c.JSON(http.StatusBadRequest, output)
+		param := new(SeedUrlextendParam)
+		if err := cc.BindAndValidate(param); err != nil {
+			return rumerrors.NewBadRequestError(err)
 		}
 
-		seed, _, err := handlers.UrlToGroupSeed(input["seed"])
+		seed, _, err := handlers.UrlToGroupSeed(param.SeedURL)
 		if err != nil {
-			output[ERROR_INFO] = err.Error()
-			return c.JSON(http.StatusBadRequest, output)
+			return rumerrors.NewBadRequestError(err)
 		}
 
 		return c.JSON(http.StatusOK, seed)
