+++ pkg/nodesdk/api/getgroupinfo.go
@@ -22,7 +22,6 @@ func (h *NodeSDKHandler) GetGroupInfo(c echo.Context) (err error) {
 
 	reqItem := new(GrpInfo)
 	reqItem.GroupId = groupid
-	reqItem.JwtToken = JwtToken
 
 	itemBytes, err := json.Marshal(reqItem)
 	if err != nil {
