+++ pkg/nodesdk/api/server.go
@@ -31,6 +31,7 @@ func StartNodeSDKServer(config StartAPIParam, signalch chan os.Signal, h *NodeSD
 	r.POST("/v1/group/getctn", h.GetGroupCtn())
 	r.POST("/v1/group/profile", h.UpdProfile)
 	r.POST("/v1/group/apihosts", h.UpdApiHostUrl)
+	r.GET("/v1/group/:group_id/apihosts", h.GetApiHostUrl)
 	r.POST("/v1/keystore/create", h.CreateNewKeyWithAlias())
 	r.POST("/v1/keystore/bindalias", h.BindAliasWithKeyName())
 	r.POST("/v1/keystore/remove", h.RmAlias())
