+++ pkg/nodesdk/http/httpclient.go
@@ -9,8 +9,8 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/rumsystem/quorum/cmd/cli/config"
 	"github.com/rumsystem/quorum/internal/pkg/logging"
+	"github.com/rumsystem/quorum/internal/pkg/utils"
 	u2 "github.com/rumsystem/quorum/internal/pkg/utils"
 )
 
@@ -18,6 +18,7 @@ var http_log = logging.Logger("http")
 
 type APIServerItem struct {
 	url      string
+	jwt      string
 	pinginms int
 	memo     string
 }
@@ -34,26 +35,24 @@ func (hc *HttpClient) Init() error {
 func (hc *HttpClient) UpdApiServer(urls []string) error {
 	http_log.Infof("UpdApiServer called")
 
-	/*
-		if len(urls) == 0 {
-			return errors.New("At least 1 url should be provided")
-		}
-	*/
-
 	var apis []*APIServerItem
 
 	for _, u := range urls {
-		_, err := url.Parse(u)
+		_url, jwt, err := utils.ParseChainapiURL(u)
 		if err != nil {
 			return errors.New("Invalid url")
 		}
+		if jwt == "" {
+			return errors.New("Invalid jwt")
+		}
 
 		var urlItem *APIServerItem
-		urlItem = &APIServerItem{}
-
-		urlItem.url = u
-		urlItem.pinginms = 0
-		urlItem.memo = ""
+		urlItem = &APIServerItem{
+			url:      _url,
+			jwt:      jwt,
+			pinginms: 0,
+			memo:     "",
+		}
 		apis = append(apis, urlItem)
 	}
 	//set group API
@@ -65,12 +64,11 @@ func (hc *HttpClient) UpdApiServer(urls []string) error {
 func (hc *HttpClient) Get(url string) ([]byte, error) {
 	http_log.Infof("Get called, groupId <%s>", url)
 
-	fullUrl, err := hc.getFullUrl(url)
+	fullUrl, jwt, err := hc.getFullUrl(url)
 	if err != nil {
 		return nil, err
 	}
 
-	jwt := config.RumConfig.Quorum.JWT
 	client, err := u2.NewHTTPClient() // hc.newHTTPClient()
 	if err != nil {
 		return nil, err
@@ -99,12 +97,11 @@ func (hc *HttpClient) Get(url string) ([]byte, error) {
 func (hc *HttpClient) GetWithBody(url string, reqData []byte) ([]byte, error) {
 	http_log.Infof("Get called, groupId <%s>", url)
 
-	fullUrl, err := hc.getFullUrl(url)
+	fullUrl, jwt, err := hc.getFullUrl(url)
 	if err != nil {
 		return nil, err
 	}
 
-	jwt := config.RumConfig.Quorum.JWT
 	client, err := u2.NewHTTPClient() // hc.newHTTPClient()
 	if err != nil {
 		return nil, err
@@ -133,23 +130,23 @@ func (hc *HttpClient) GetWithBody(url string, reqData []byte) ([]byte, error) {
 func (hc *HttpClient) Post(url string, data []byte) ([]byte, error) {
 	http_log.Infof("Post called, <%s>", url)
 
-	fullUrl, err := hc.getFullUrl(url)
+	fullUrl, jwt, err := hc.getFullUrl(url)
 	if err != nil {
 		return nil, err
 	}
 
-	jwt := config.RumConfig.Quorum.JWT
 	client, err := u2.NewHTTPClient() // hc.newHTTPClient()
 	if err != nil {
 		return nil, err
 	}
 	req, err := http.NewRequest(http.MethodPost, fullUrl, bytes.NewBuffer(data))
+	if err != nil {
+		return nil, err
+	}
+
 	req.Header.Set("Content-Type", "application/json; charset=utf-8")
 	if jwt != "" {
 		req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", jwt))
-	}
-	if err != nil {
-		return nil, err
 	}
 
 	resp, err := client.Do(req)
@@ -169,12 +166,11 @@ func (hc *HttpClient) Post(url string, data []byte) ([]byte, error) {
 func (hc *HttpClient) Delete(url string, data []byte) ([]byte, error) {
 	http_log.Infof("Delete called %s", url)
 
-	fullUrl, err := hc.getFullUrl(url)
+	fullUrl, jwt, err := hc.getFullUrl(url)
 	if err != nil {
 		return nil, err
 	}
 
-	jwt := config.RumConfig.Quorum.JWT
 	client, err := u2.NewHTTPClient() //hc.newHTTPClient()
 	if err != nil {
 		return nil, err
@@ -200,20 +196,23 @@ func (hc *HttpClient) Delete(url string, data []byte) ([]byte, error) {
 	return body, err
 }
 
-func (hc *HttpClient) getFullUrl(u2 string) (fullurl string, err error) {
+func (hc *HttpClient) getFullUrl(path string) (fullurl string, jwt string, err error) {
 	//TBD: find the fastest api
 	//TBD: skip unavailable api
 	//now just return the first URL of remote api in the list
-	result := hc.APIs[0].url + u2
+	jwt = hc.APIs[0].jwt
+	baseUrl := hc.APIs[0].url
 
-	http_log.Infof("url %s", result)
-
-	_, err = url.Parse(result)
+	u, err := url.Parse(baseUrl)
 	if err != nil {
-		return "", errors.New("Can not get Full Url, url invalid")
+		return "", "", errors.New("Can not get Full Url, url invalid")
 	}
+	u.Path = path
+	fullurl = u.String()
 
-	return result, nil
+	http_log.Debugf("fullurl: %s", fullurl)
+
+	return fullurl, jwt, nil
 }
 
 func (hc *HttpClient) checkJWTError(body string) error {
